Update the section 1.1 Triangle Soup Editor
### *1.1 Triangle Soup Editor*

key "i": Triangle Insertion Mode
* ENABLED: Every triple of subsequent mouse clicks will create a triangle. 
The first click will create the starting point of the segment, which will be immediately visualized. 
As the mouse is moved, a preview of a segment will appear. 
After the second mouse click, a preview of the triangle will appear. 
After the third click, the current preview will transform into the final triangle.

key "o": Triangle Translation Mode
* ENABLED: Each mouse click will selected the triangle below the cursor (which will be highlighted), 
and every movement of the mouse (while keeping the button pressed) will result in a corresponding translation of the triangle. 

key "p": Triangle Deletion Mode
*  ENABLED: When the mouse is pressed, the triangle below the cursor is deleted.